+++ scripts/debt-ratchet-baseline.json
@@ -4,7 +4,7 @@
   "counts": {
     "any_casts": 0,
     "todo_comments": 0,
-    "oversized_files": 24,
+    "oversized_files": 16,
     "legacy_wordpress_runtime": 0,
     "bootstrap_imports": 0,
     "redux_imports": 0,
@@ -13,7 +13,6 @@
   "oversized_file_allowlist": [
     "components/about/AboutTermsOfService.tsx",
     "components/auth/SeizeConnectProvider.tsx",
-    "components/brain/left-sidebar/waves/memes-quick-vote/useMemesQuickVotePreviewSwipe.ts",
     "components/distribution-plan-tool/build-phases/build-phase/form/BuildPhaseFormConfigModal.tsx",
     "components/distribution-plan-tool/review-distribution-plan/table/ReviewDistributionPlanTableSubscriptionFooter.tsx",
     "components/drops/view/item/content/attachments/DropAttachmentDisplay.tsx",
@@ -24,16 +23,9 @@
     "components/memelab/MemeLabPage.tsx",
     "components/nextGen/admin/NextGenAdmin.tsx",
     "components/nft-transfer/TransferModal.tsx",
-    "components/notifications/NotificationsContext.tsx",
     "components/react-query-wrapper/ReactQueryWrapper.tsx",
     "components/rememes/RememePage.tsx",
     "components/user/collected/UserPageCollected.tsx",
-    "components/waves/WaveTrustSignals.tsx",
-    "components/waves/drops/WaveDrop.tsx",
-    "components/waves/drops/WaveDropReactions.tsx",
-    "contexts/wave/hooks/useWaveRealtimeUpdater.ts",
-    "helpers/Helpers.ts",
-    "hooks/useMemesQuickVoteQueue.ts",
-    "services/api/wave-drops-v2-api.ts"
+    "helpers/Helpers.ts"
   ]
 }
